// frontend/src/app/shared/models/creatures/monsters/monster-type.enum.ts
export enum MonsterType {
  ABERRATION = 'ABERRATION',
  BEAST = 'BEAST',
  CELESTIAL = 'CELESTIAL',
  CONSTRUCT = 'CONSTRUCT',
  DRAGON = 'DRAGON',
  ELEMENTAL = 'ELEMENTAL',
  FEY = 'FEY',
  FIEND = 'FIEND',
  GIANT = 'GIANT',
  HUMANOID = 'HUMANOID',
  MONSTROSITY = 'MONSTROSITY',
  OOZE = 'OOZE',
  PLANT = 'PLANT',
  UNDEAD = 'UNDEAD'
}